Move test fixture class to upper level
